backend/requirements.txt
# Core Flask and web framework
Flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.3

# Database and ORM
SQLAlchemy==2.0.32
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9

# Authentication and security
flask-jwt-extended==4.6.0
python-dotenv==1.0.1

# PDF processing
PyPDF2==3.0.1
pymupdf==1.24.9

# Machine Learning and NLP
transformers==4.43.3
torch==2.5.0
scikit-learn==1.5.1
nltk==3.8.1

# API and web requests
requests==2.32.3

# Data validation and serialization
marshmallow==3.21.3

# Google APIs (optional)
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# Report generation
reportlab==4.2.2

# Development and testing
pytest==8.3.2
pytest-flask==1.3.0
coverage==7.6.1